Refactoring Parser
